Reorg repeated feedback
